Added JavaDoc to most functions and classes. Moved getInput from ui package to parser package. Removed exception handler when registering to-do task.
