+++ src/main/java/duke/util/TaskList.java
@@ -4,6 +4,10 @@ import duke.task.Task;
 
 import java.util.ArrayList;
 
+/**
+ * A <code>TaskList</code> object contains the ArrayList of tasks from
+ * user inputs. Handles the deletion of items and searching of list.
+ */
 public class TaskList {
     public static ArrayList<Task> lists;
     private Ui ui;
@@ -12,7 +16,10 @@ public class TaskList {
         this.lists = lists;
     }
 
-
+    /**
+     * Lists all task recorded in the given list.
+     *
+     */
     public void listTasks() {
         System.out.println("Here are the tasks in your list:");
         int itemNumber = 1;
@@ -25,6 +32,13 @@ public class TaskList {
         }
     }
 
+    /**
+     * Marks/unmarks or calls for deletion of task at the specified position.
+     *
+     * @param words Divided user input - to get item number of task.
+     * @param line User input.
+     * @throws IndexOutOfBoundsException If item number provided does not exist in ArrayList.
+     */
     public void markOrDeleteTask(String[] words, String line) {
         try {
             int itemNumber = Integer.parseInt(words[1]);
@@ -41,6 +55,11 @@ public class TaskList {
         }
     }
 
+    /**
+     * Deletes task at the specified position.
+     *
+     * @param itemNumber Item number of task for deletion.
+     */
     public void DeleteTask(int itemNumber) {
         int index = itemNumber - 1;
         Task item = lists.get(index);
